metis.readgcam

This function connects to a gcamdatabase and uses a query file to out results into a table ready for plotting.

metis.readgcam(gcamdatabasePath = NULL, gcamdatabaseName = NULL,
  queryxml = "metisQueries.xml", queryPath = paste(getwd(),
  "/dataFiles/gcam", sep = ""), scenOrigNames = NULL,
  scenNewNames = NULL, reReadData = T, dataProj = "dataProj.proj",
  dataProjPath = paste(getwd(), "/outputs", sep = ""),
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  regionsSelect = NULL, queriesSelect = "All", paramsSelect = "All",
  folderName = NULL, nameAppend = "")

Arguments

gcamdatabasePath
Path to gcam database folder

gcamdatabaseName
Name of gcam database

queryxml
Name of the query.xml file. By default it is "metisQueries.xml"

queryPath
Folder that contains the query.xml file.By default it is the same folder as specified by gcamdatabasePath

scenOrigNames
Original Scenarios names in GCAM database in a string vector. For example c('scenario1','scenario2).

scenNewNames
New Names which may be shorter and more useful for figures etc. Default will use Original Names. For example c('scenario1','scenario2)

reReadData
If TRUE will read the GCAM data base and create a queryData.proj file in the same folder as the GCAM database. If FALSE will load a '.proj' file if a file with full path is provided otherwise it will search for a dataProj.proj file in the existing folder which may have been created from an old run.

dataProj
Optional. A default 'dataProj.proj' is produced if no .Proj file is specified.

dataProjPath
Folder that contains the dataProj or where it will be produced. By default it is the same folder as specified by gcamdatabasePath

dirOutputs
Full path to directory for outputs

regionsSelect
The regions to analyze in a vector. Example c('Colombia','Argentina'). Full list: c(USA, Africa_Eastern, Africa_Northern, Africa_Southern, Africa_Western, Australia_NZ, Brazil, Canada Central America and Caribbean, Central Asia, China, EU-12, EU-15, Europe_Eastern, Europe_Non_EU, European Free Trade Association, India, Indonesia, Japan, Mexico, Middle East, Pakistan, Russia, South Africa, South America_Northern, South America_Southern, South Asia, South Korea, Southeast Asia,

queriesSelect
Default = "All". Predetermined subsets or a vector of queries to read from the queryxml for example predetermined subsets would be c('water','energy') or selection of queries would be c("Total final energy by aggregate end-use sector", "Population by region"). The queries must be availble in the queryxml file. Queryset names include: c("water", "energy", "land", "emissions", "ag", "socioecon", "transport") Current list of queries for each set include: water

"water withdrawals by crop"

"water withdrawals by water mapping source"

"water consumption by water mapping source"

"water withdrawals by sector"

"water consumption by sector"

"biophysical water demand by crop type and land region"

"Basin level available runoff"

"water withdrawals by state, sector, basin (includes desal)"

"water consumption by state, sector, basin (includes desal)"

energy

"primary energy consumption by region (direct equivalent) ORDERED SUBSECTORS"

"Final energy by detailed end-use sector and fuel"

"total final energy by aggregate sector"

"refined liquids production by subsector"

"building final energy by fuel"

"industry final energy by fuel"

"building final energy by subsector"

"transport final energy by fuel"

"transport final energy by mode and fuel"

electricity

"elec gen by gen tech USA"

"elec gen by gen tech cogen USA"

"elec gen by gen tech and cooling tech"

"Electricity generation by aggregate technology"

land

"land allocation by crop and water source",

"aggregated land allocation",

"land allocation by crop"

emissions

"nonCO2 emissions by resource production",

"nonCO2 emissions by sector"

"Land Use Change Emission (future)"

"CO2 emissions by sector (no bio)"

"CO2 emissions by sector"

ag

"Ag Production by Crop Type"

"ag production by tech"

socioecon

"GDP MER by region"

"GDP per capita MER by region"

"Population by region"

transport

"transport service output by mode"

"transport service output by tech (new)"

paramsSelect
Default = "All". If desired dplyr::select a subset of paramaters to analyze from the full list of parameters: c(# Energy "energyPrimaryByFuelEJ","energyPrimaryRefLiqProdEJ", "energyFinalConsumBySecEJ","energyFinalByFuelBySectorEJ","energyFinalSubsecByFuelTranspEJ", "energyFinalSubsecByFuelBuildEJ", "energyFinalSubsecByFuelIndusEJ","energyFinalSubsecBySectorBuildEJ", "energyPrimaryByFuelMTOE","energyPrimaryRefLiqProdMTOE", "energyFinalConsumBySecMTOE","energyFinalbyFuelMTOE","energyFinalSubsecByFuelTranspMTOE", "energyFinalSubsecByFuelBuildMTOE", "energyFinalSubsecByFuelIndusMTOE","energyFinalSubsecBySectorBuildMTOE", "energyPrimaryByFuelTWh","energyPrimaryRefLiqProdTWh", "energyFinalConsumBySecTWh","energyFinalbyFuelTWh","energyFinalSubsecByFuelTranspTWh", "energyFinalSubsecByFuelBuildTWh", "energyFinalSubsecByFuelIndusTWh","energyFinalSubsecBySectorBuildTWh", # Electricity "elecByTechTWh","elecCapByFuel","elecFinalBySecTWh","elecFinalByFuelTWh", "elecNewCapCost","elecNewCapGW","elecAnnualRetPrematureCost","elecAnnualRetPrematureGW","elecCumCapCost","elecCumCapGW","elecCumRetPrematureCost","elecCumRetPrematureGW", # Transport "transportPassengerVMTByMode", "transportFreightVMTByMode", "transportPassengerVMTByFuel", "transportFreightVMTByFuel", # Water "watConsumBySec", "watWithdrawBySec", "watWithdrawByCrop", "watBioPhysCons", "watIrrWithdrawBasin","watIrrConsBasin", # Socio-economics "gdpPerCapita", "gdp", "gdpGrowthRate", "pop", "livestock_MeatDairybyTechMixed","livestock_MeatDairybyTechPastoral","livestock_MeatDairybyTechImports", "livestock_MeatDairybySubsector", #Land use "landIrrRfd", "landIrrCrop","landRfdCrop", "landAlloc","landAllocByCrop",# Emissions "emissLUC", "emissNonCO2BySectorGWPAR5","emissNonCO2BySectorGTPAR5","emissNonCO2BySectorOrigUnits", "emissNonCO2ByResProdGWPAR5", "emissBySectorGWPAR5FFI","emissMethaneBySourceGWPAR5", "emissByGasGWPAR5FFI", "emissByGasGWPAR5LUC", "emissBySectorGWPAR5LUC", "emissNonCO2ByResProdGTPAR5", "emissBySectorGTPAR5FFI","emissMethaneBySourceGTPAR5", "emissByGasGTPAR5FFI", "emissByGasGTPAR5LUC","emissBySectorGTPAR5LUC", "emissCO2BySectorNoBio" #' @return A list with the scenarios in the gcam database, queries in the queryxml file and a tibble with gcam data formatted for metis charts.

folderName
Default = NULL

nameAppend
Default="". Name to append to saved files.